requirements.txt
attrs==20.2.0
bcrypt==3.2.0
blinker==1.4
cffi==1.14.3
click==7.1.2
Flask==1.1.2
Flask-Admin==1.5.7
Flask-Bcrypt==0.7.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.4
Flask-Testing==0.8.0
importlib-metadata==2.0.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.4
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.9.0
pycparser==2.20
pyparsing==2.4.7
pytest==6.1.1
python-dotenv==0.14.0
redis==3.5.3
rq==1.5.2
six==1.15.0
SQLAlchemy==1.3.20
toml==0.10.1
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.4.0
